Small refactoring
